backend/app/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic-settings==2.6.1
pydantic_core==2.41.5
python-dotenv==1.2.2
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
psycopg2-binary==2.9.9
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.7
email-validator==2.2.0
bcrypt==4.2.1
cryptography==44.0.0
# Google Calendar + Meet integration
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
